Rename and reparameterize SampleDeleteConfirmModal to EntityDeleteConfirmModal (and move to lineage instead of samples) for reuse with data class objects
